edit forumcontroller, topic.html

+++ src/main/java/forum/service/TopicServiceJPA.java
@@ -27,11 +27,11 @@ public class TopicServiceJPA {
 				.setParameter("ident", ident).executeUpdate();
 	}
 	
-	public int getTopicCountForCategory(int ident) {
-		return (int) entityManager.createQuery("SELECT COUNT(t) FROM Topic t WHERE t.category_id = :ident").setParameter("ident", ident).getSingleResult();
+	public long getTopicCountForCategory(int ident) {
+		return (long) entityManager.createQuery("SELECT COUNT(t) FROM Topic t WHERE t.category_id = :ident").setParameter("ident", ident).getSingleResult();
 	}
 	
-	public int getTopicCount() {
-		return (int) entityManager.createQuery("SELECT COUNT(t) FROM Topic t").getSingleResult();
+	public long getTopicCount() {
+		return (long) entityManager.createQuery("SELECT COUNT(t) FROM Topic t").getSingleResult();
 	}
 }
